[#24] Optimize DefaultByteArrayNodeFactory for single character nodes
Adds six more implementation classes to
com.googlecode.concurrenttrees.radix.node.concrete.bytearray which
optimize the case when a node has an incoming edge of a single character
only. Instead of storing it in a byte array a simple byte field is used.
This significantly lowers the memory overhead of such nodes.

TODO:
- Dedicated unit tests.